修改金额相关字段为bigint  防止前端输入较大金额时的数据库报错问题

+++ database/migrations/2017_05_22_043802_wallets.php
@@ -27,7 +27,7 @@ class Wallets extends Migration
 
             // 余额，单位 分 (避免小数计算偏移)
             $table
-                ->integer('balance')
+                ->bigInteger('balance')
                 ->unsigned()
                 ->comment('钱包余额');
             $table->timestamps(); // 自动维护时间
