+++ database/migrations/2017_06_06_094947_wallet_charges.php
@@ -46,7 +46,7 @@ class WalletCharges extends Migration
                 ->comment('类型：1 - 增加，0 - 减少');
 
             $table
-                ->integer('amount')
+                ->bigInteger('amount')
                 ->unsigned()
                 ->comment('总额');
